VRAG_s3/requirements.txt
psutil==7.0.0
colorama==0.4.6
verl==0.2.0.dev0
attrs==25.3.0
sniffio==1.3.1
proto-plus==1.26.1
narwhals==2.3.0
numba==0.60.0
mypy_extensions==1.1.0
colpali_engine==0.3.11.dev7+g481eb2008
scipy==1.15.3
nest-asyncio==1.6.0
joblib==1.5.2
httpx==0.28.1
websocket-client==1.8.0
sentence-transformers==5.1.0
nvidia-nvtx-cu12==12.4.127
pybind11==3.0.1
transformers==4.57.0
PyYAML==6.0.2
pip==25.2
fastapi-cloud-cli==0.1.5
pydantic-extra-types==2.10.5
starlette==0.47.3
mpmath==1.3.0
setuptools==78.1.1
llama-index-embeddings-huggingface==0.4.0
protobuf==4.25.8
watchfiles==1.1.0
outlines_core==0.1.26
urllib3==2.5.0
streamlit==1.49.1
pyvers==0.1.0
cupy-cuda12x==13.6.0
torchdata==0.11.0
Deprecated==1.2.18
llvmlite==0.43.0
pycountry==24.6.1
nvidia-cudnn-cu12==9.1.0.70
prometheus-fastapi-instrumentator==7.1.0
mistral_common==1.8.4
blinker==1.9.0
huggingface-hub==0.34.4
nvidia-curand-cu12==10.3.5.147
aiohappyeyeballs==2.6.1
bitsandbytes==0.47.0
googleapis-common-protos==1.70.0
accelerate==1.10.1
openai==1.105.0
opencv-python-headless==4.11.0.86
flash_attn==2.8.3
tiktoken==0.11.0
torchvision==0.21.0
propcache==0.3.2
fsspec==2025.3.0
psutil==7.0.0
dnspython==2.7.0
hf-xet==1.1.9
tensordict==0.6.0
llama-index-question-gen-openai==0.3.0
llama-index-llms-openai==0.3.1
triton==3.2.0
toml==0.10.2
scikit-learn==1.7.1
llama-index-multi-modal-llms-openai==0.3.0
dataclasses-json==0.6.7
xxhash==3.5.0
gitdb==4.0.12
soupsieve==2.8
google-api-core==2.25.1
fastapi-cli==0.0.10
aiohttp==3.12.15
fastrlock==0.8.3
llama-index-indices-managed-llama-cloud==0.6.2
grpcio==1.74.0
tqdm==4.67.1
platformdirs==4.4.0
sentry-sdk==2.35.2
py-spy==0.4.1
orjson==3.11.3
llama-index-embeddings-openai==0.3.0
Pygments==2.19.2
gguf==0.10.0
MarkupSafe==3.0.2
tornado==6.5.2
packaging==25.0
torchaudio==2.6.0
astor==0.8.1
pylatexenc==2.10
rignore==0.6.4
certifi==2025.8.3
lm-format-enforcer==0.10.12
opentelemetry-api==1.36.0
multiprocess==0.70.16
sentencepiece==0.2.1
opentelemetry-sdk==1.36.0
pyasn1==0.6.1
llama-index-legacy==0.9.48.post4
aiosignal==1.4.0
google-auth==2.40.3
tenacity==8.5.0
nvidia-cublas-cu12==12.4.5.8
idna==3.10
jsonschema==4.25.1
pypdf==5.9.0
torch==2.6.0
outlines==0.1.11
wheel==0.45.1
opencensus-context==0.1.3
mdurl==0.1.2
typer==0.17.3
aiohttp-cors==0.8.1
nvidia-nccl-cu12==2.21.5
cachetools==5.5.2
nvidia-cusparselt-cu12==0.6.2
SQLAlchemy==2.0.43
filelock==3.19.1
smart_open==7.3.0.post1
opencensus==0.11.4
tzdata==2025.2
safetensors==0.6.2
rich-toolkit==0.15.0
ninja==1.13.0
Werkzeug==3.1.3
regex==2025.9.1
llama-cloud==0.1.5
distro==1.9.0
httptools==0.6.4
airportsdata==20250811
marshmallow==3.26.1
email-validator==2.3.0
pycparser==2.23
wandb==0.21.3
llama-index-core==0.12.1
py-cpuinfo==9.0.0
cffi==2.0.0
msgspec==0.19.0
verl==0.2.0.dev0
opentelemetry-exporter-prometheus==0.57b0
interegular==0.3.3
opentelemetry-semantic-conventions==0.57b0
itsdangerous==2.2.0
nvidia-cufft-cu12==11.2.1.3
prometheus_client==0.23.1
charset-normalizer==3.4.3
pyarrow==21.0.0
pydantic_core==2.23.4
dashscope==1.24.5
nvidia-cuda-runtime-cu12==12.4.127
watchdog==6.0.0
GitPython==3.1.45
virtualenv==20.34.0
nvidia-nvjitlink-cu12==12.4.127
async-timeout==5.0.1
vllm==0.8.2
beautifulsoup4==4.13.5
xformers==0.0.29.post2
partial-json-parser==0.2.1.1.post6
six==1.17.0
nvidia-cuda-cupti-cu12==12.4.127
threadpoolctl==3.6.0
dill==0.3.8
wrapt==1.17.3
blake3==1.0.5
httpcore==1.0.9
fastapi==0.116.1
peft==0.15.2
llama-parse==0.5.14
altair==5.5.0
typing-inspect==0.9.0
nvidia-cusolver-cu12==11.6.1.9
diskcache==5.6.3
sympy==1.13.1
jiter==0.10.0
einops==0.8.1
opentelemetry-proto==1.36.0
Jinja2==3.1.6
latex2sympy2_extended==1.10.2
greenlet==3.2.4
distlib==0.4.0
python-dateutil==2.9.0.post0
nvidia-cuda-nvrtc-cu12==12.4.127
python-json-logger==3.3.0
python-multipart==0.0.20
yarl==1.20.1
dirtyjson==1.0.8
frozenlist==1.7.0
filetype==1.2.0
llama-index==0.12.0
llama-index-readers-llama-parse==0.4.0
cryptography==46.0.1
markdown-it-py==4.0.0
typing_extensions==4.15.0
click==8.3.0
codetiming==1.4.0
llama-index-agent-openai==0.4.0
lark==1.2.2
pydantic==2.9.2
pytz==2025.2
rpds-py==0.27.1
omegaconf==2.3.0
h11==0.16.0
uvicorn==0.35.0
smmap==5.0.2
ray==2.49.2
msgpack==1.1.1
rich==14.1.0
llama-index-program-openai==0.3.0
math-verify==0.8.0
requests==2.32.5
hydra-core==1.3.2
referencing==0.36.2
pillow==11.3.0
llama-index-cli==0.4.0
nvidia-cusparse-cu12==12.3.1.170
annotated-types==0.7.0
pydeck==0.9.1
networkx==3.4.2
websockets==15.0.1
numpy==1.26.4
jsonschema-specifications==2025.9.1
exceptiongroup==1.3.0
striprtf==0.0.26
anyio==4.10.0
pandas==2.3.2
uvloop==0.21.0
compressed-tensors==0.9.2
antlr4-python3-runtime==4.9.3
depyf==0.18.0
Flask==3.1.2
colorful==0.5.7
nltk==3.9.1
llama-index-readers-file==0.4.0
zipp==3.23.0
python-dotenv==1.1.1
xgrammar==0.1.16
datasets==4.0.0
cloudpickle==3.1.1
shellingham==1.5.4
rsa==4.9.1
importlib_metadata==8.7.0
tokenizers==0.22.1
llguidance==0.7.30
pyasn1_modules==0.4.2
liger_kernel==0.6.2
nvidia-ml-py==13.580.82
multidict==6.6.4
pyzmq==27.0.2
inflect==7.3.1
jaraco.context==5.3.0
wheel==0.45.1
jaraco.collections==5.1.0
platformdirs==4.2.2
jaraco.text==3.12.1
tomli==2.0.1
importlib_metadata==8.0.0
autocommand==2.2.2
zipp==3.19.2
typeguard==4.3.0
backports.tarfile==1.2.0
jaraco.functools==4.0.1
more-itertools==10.3.0
packaging==24.2
typing_extensions==4.12.2
